refactor : recommanded  이전
